+++ src/main/java/net/neganote/gtutilities/datagen/lang/UtilLangHandler.java
@@ -23,6 +23,9 @@ public class UtilLangHandler {
         provider.add("tooltip.pterb_machine.input_coolant_before_use", "Always input coolant before turning it on!");
         provider.add("tooltip.pterb_machine.dont_use_terminal", "DO NOT BUILD WITH THE TERMINAL, VERY BUGGY");
 
+        provider.add("gtmutils.machine.64a_energy_converter.tooltip_tool_usage",
+                "Starts as §fEU Converter§7, use Soft Mallet to change");
+
         provider.add("gtmutils.multiblock.pterb_machine.coolant_usage", "§cDrains %sL of %s per second");
         provider.add("gtmutils.pterb_machine.invalid_frequency", "PTERBs will not work on frequency 0!");