+++ changelog.md
@@ -1,3 +1,3 @@
 # GregTech Modern Utilities version 1.1.2
 
-HOTFIX: fix yet another PTERB crash, sorry
+HOTFIX: fix yet another PTERB crash as well as a structure issue, sorry