+++ changelog.md
@@ -1,3 +1,3 @@
-# GregTech Modern Utilities version 1.1.5
+# GregTech Modern Utilities version 1.1.6
 
-* Fix various PTERB issues
+* Fix PTERBs that aren't in the overworld not sending power