+++ changelog.md
@@ -1,3 +1,3 @@
-# GregTech Modern Utilities version 1.1.7
+# GregTech Modern Utilities version 1.1.8
 
-* Change PTERB structure to use superconducting coils, force limit on amount of energy/laser hatches
+* Fix tooltip to reflect 64a converters starting as EU in