+++ changelog.md
@@ -1,3 +1,3 @@
-# GregTech Modern Utilities version 1.1.11
+# GregTech Modern Utilities version 2.0.0
 
 * Fix terminal problems (as much as possible) and remove warning about building PTERB with terminal