+++ changelog.md
@@ -1,3 +1,4 @@
 # GregTech Modern Utilities version 2.0.0
 
-* Fix terminal problems (as much as possible) and remove warning about building PTERB with terminal
+* Port to GTM 7.0.0
+* Make omni-breaker teleport drops to player on right-click